# apps/api/pkg/swagger/swagger.yaml
basePath: /
definitions:
  common.CommonErrorSchema:
    properties:
      error:
        type: string
    type: object
  common.CommonPaginationSchema:
    properties:
      hasNext:
        type: boolean
      page:
        type: integer
      pageSize:
        type: integer
      totalCount:
        type: integer
    type: object
  entities.User:
    properties:
      active:
        type: boolean
      avatar:
        type: string
      bio:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      subDragonId:
        type: integer
      updatedAt:
        type: string
      username:
        type: string
    required:
    - id
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  users.UpdateUserSchema:
    properties:
      avatar:
        type: string
      bio:
        type: string
      email:
        type: string
      subDragonId:
        type: integer
      username:
        maxLength: 30
        minLength: 3
        type: string
    type: object
  users.UserListSchema:
    properties:
      data:
        items:
          $ref: '#/definitions/entities.User'
        type: array
      pagination:
        $ref: '#/definitions/common.CommonPaginationSchema'
    type: object
  users.UserSchema:
    properties:
      avatar:
        type: string
      bio:
        type: string
      email:
        type: string
      subDragonId:
        type: integer
      username:
        maxLength: 30
        minLength: 3
        type: string
    required:
    - email
    - username
    type: object
host: localhost:3333
info:
  contact:
    email: fiber@swagger.io
    name: API Support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
  /v1/users:
    get:
      consumes:
      - application/json
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of items per page (default: 10)'
        in: query
        name: pageSize
        type: integer
      - description: Get users by field (optional)
        enum:
        - sub_dragon_id
        in: query
        name: getBy
        type: string
      - description: Get users by field value (optional)
        in: query
        name: getByValue
        type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/users.UserListSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.CommonErrorSchema'
      summary: Get User List
      tags:
      - User
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.UserSchema'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.CommonErrorSchema'
      summary: Create User
      tags:
      - User
  /v1/users/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body for updating user
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.UpdateUserSchema'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.CommonErrorSchema'
      summary: Update User
      tags:
      - User
  /v1/users/{userId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userId
        type: integer
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.CommonErrorSchema'
        "404":
          description: Not Found Request
          schema:
            $ref: '#/definitions/common.CommonErrorSchema'
      summary: Delete user by ID
      tags:
      - User
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userId
        type: integer
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/users.UserSchema'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.CommonErrorSchema'
        "404":
          description: Not Found Request
          schema:
            $ref: '#/definitions/common.CommonErrorSchema'
      summary: Get user by ID
      tags:
      - User
swagger: "2.0"